Fixed nav and footer repetition - replaced px with rem on homepage
