+++ blog-page.css
@@ -1,20 +1,4 @@
-@import url('https://fonts.googleapis.com/css2?family=DM+Serif+Text&family=Poppins:wght@100;200;300;400&display=swap');
 
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-
-    font-family: 'Poppins', sans-serif;
-}
-
-body {
-    /* border: 2px solid red; */
-
-    --purple: #58287F;
-    --green: #0B9444;
-    --chocolate: #1A0000;
-}
 
 /* <!-- UNIVERSAL --> */
