+++ blogs.html
@@ -14,65 +14,6 @@
 
     <div class="hero-nav">
 
-        <!-- NAVBAR  -->
-
-        <div class="navbar">
-
-            <div class="logo">
-                <a href="./index.html">
-                    <img src="images/logo.png" alt="Logo">
-                </a>
-            </div>
-
-            <div class="nav-links">
-                <ul>
-                    <li><a href="./index.html#about">About</a></li>
-                    <li><a href="./blogs.html">Blog</a></li>
-                    <li><a href="#footer">Contact</a></li>
-                </ul>
-            </div>
-
-            <div class="btn-primary navbar-cta">
-                <a href="./index.html#portfolio">Portfolio</a>
-            </div>
-
-            <div class="hamburger responsive">
-                <a class="hamburger-icon" href="#">
-                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5"
-                        stroke="currentColor" class="w-6 h-6">
-                        <path stroke-linecap="round" stroke-linejoin="round"
-                            d="M3.75 6.75h16.5M3.75 12h16.5m-16.5 5.25h16.5" />
-                    </svg>
-
-                </a>
-                <a class="hamburger-icon is-not-active" href="#">
-                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5"
-                        stroke="currentColor" class="w-6 h-6">
-                        <path stroke-linecap="round" stroke-linejoin="round"
-                            d="M9.75 9.75l4.5 4.5m0-4.5l-4.5 4.5M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
-                    </svg>
-
-                </a>
-            </div>
-
-        </div>
-
-        <div class="responsive-nav hide-nav">
-
-            <div class="nav-links">
-                <ul>
-                    <li><a href="./index.html#about">About</a></li>
-                    <li><a href="./blogs.html">Blog</a></li>
-                    <li><a href="#footer">Contact</a></li>
-                </ul>
-            </div>
-
-            <div class="navbar-cta">
-                <a class="btn-primary navbar-cta" href="./index.html#portfolio">Portfolio</a>
-            </div>
-
-        </div>
-
         <!-- HERO SECTION  -->
 
         <div class="blogs-hero">
@@ -132,164 +73,7 @@
         
         </div>
 
-    </div>
-
-    <!-- FOOTER  -->
-
-    <div class="footer">
-
-        <!-- FOOTER UP -->
-
-        <div class="footer-up">
-
-            <span>
-                <div class="quick-links">
-
-                    <h1>
-                        Quick Links
-                    </h1>
-    
-                    <ul>
-    
-                        <li>
-                            <a href="./index.html#about">
-                                About
-                            </a>
-                        </li>
-    
-                        <li>
-                            <a href="./blogs.html">
-                                Blog
-                            </a>
-                        </li>
-    
-                        <li>
-                            <a href="#footer">
-                                Contact
-                            </a>
-                        </li>
-    
-                        <li>
-                            <a id="admin-link" href="./login.html">
-                                Admin
-                            </a>
-                        </li>
-    
-                    </ul>
-    
-                </div>
-
-                <div class="newsletter">
-                        
-                        <h1>
-                            Newsletter
-                        </h1>
-    
-                        <p>
-                            Subscribe to our newsletter and get our latest news and updates.
-                        </p>
-    
-                        <form id="newsletter-form" action="">
-                            <input type="email" id="newsletter-email" placeholder="Email">
-                            <p id="newsletter-error">Please enter a valid email</p>
-                            <p id="newsletter-success">You have subscribed to our newsletter successfully!</p>
-                            <div class="form-submit">
-                                <input class="btn-primary" id="newsletter-submit" type="submit" value="Subscribe">
-                            </div>
-                        </form>
-                </div>
-            </span>
-
-            <div class="form-container" id="footer-form">
-
-                <h1>
-                    Leave a quick message
-                </h1>
-
-                <form action="">
-                    <input type="text" id="contact-form-name" placeholder="Name">
-                    <p id="contact-name-error">
-                        Please enter a valid name
-                    </p>
-                    <input type="email" id="contact-form-email" placeholder="Email">
-                    <p id="contact-email-error">
-                        Please enter a valid email
-                    </p>
-                    <textarea name="" id="contact-form-message" cols="50" rows="5" placeholder="Message"></textarea>
-                    <p id="contact-message-error">
-                        Please enter a valid message
-                    </p>
-                    <p id="contact-form-success">
-                        Thank you for reaching out!
-                        I will respond ASAP!
-                    </p>
-                    <p id="contact-form-error">
-                        Please make sure you have provided valid information
-                    </p>
-                    <div class="form-submit">
-                        <input class="btn-primary" id="contact-form-submit" type="submit" value="Send">
-                    </div>
-                </form>
-
-            </div>
-
-        </div>
-
-        <!-- FOOTER DOWN  -->
-
-        <div class="footer-down">
-
-            <div class="copyright-heading">
-
-                <h1>
-                    nishimwe<span>Prince</span>
-                </h1>
-                <p>
-                    <span>Fullstack Web Developer</span> & IT Support
-                </p>
-
-            </div>
-
-            <div class="copyright-social-icons">
-
-                <ul>
-                    <li>
-                        <a href="https://github.com/nishimweprince">
-                            <img src="./images/github-icon.png" alt="Github icon">
-                        </a>
-                    </li>
-                    <li>
-                        <a href="https://www.linkedin.com/in/nishimweprince">
-                            <img src="./images/linkedin-icon.png" alt="LinkedIn icon">
-                        </a>
-                    </li>
-                    <li>
-                        <a href="https://www.twitter.com/nishimweprince">
-                            <img src="./images/twitter-icon.png" alt="Twitter icon">
-                        </a>
-                    </li>
-                </ul>
-
-            </div>
-
-            <div class="copyright-footer">
-
-                <p>
-                    &copy; 2021 Nishimwe Prince
-                </p>
-
-                <p>
-                    All rights reserved
-                </p>
-
-            </div>
-
-        </div>
-
-    </div>
-
-<script src="index.js"></script>
-<script src="blogs.js"></script>
+<script type="module" src="blogs.js"></script>
 </body>
 
 </html>
