+++ index.css
@@ -1,0 +1,411 @@
+  /* HERO NAV  */
+
+  .hero-nav {
+    height: 100vh;
+    display: grid;
+    grid-template-rows: 15% 85%;
+}
+
+  /* HERO TEXT  */
+
+  .hero-text {
+
+    display: flex;
+    flex-direction: column;
+    gap: 40px;
+    justify-content: center;
+}
+
+.hero-text-heading {
+    display: flex;
+    flex-direction: column;
+    gap: 5px;
+}
+
+.hero-text-heading h1 {
+    font-size: 45px;
+    font-weight: 500;
+}
+
+.hero-text-heading h1 span {
+    color: #58287F;
+}
+
+.hero-text-heading p {
+    font-size: 18px;
+    font-weight: 400;
+
+    word-wrap: break-word;
+    max-width: 250px;
+}
+
+.hero-cta {
+    height: auto;
+}
+
+/* HERO IMAGE  */
+
+.hero-image {
+    background-image: url('./images/nishimwe-formal.jpg');
+    background-position: center;
+    background-size: cover;
+
+    display: flex;
+    flex-direction: column;
+    align-items: flex-end;
+    justify-content: flex-end;
+}
+
+.hero-line {
+    border-left: 3px solid #58287F;
+    height: 200px;
+    width: 2px;
+    padding: 10px;
+}
+
+.hero-line:hover {
+    transform: scale(1.05);
+    transition: all 0.3s ease-in-out;
+    cursor: pointer;
+}
+
+.hero-image ul {
+    padding: 20px 0;
+    display: flex;
+    flex-direction: column;
+    height: 40%;
+}
+
+.hero-image ul {
+    list-style: none;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-between;
+}
+
+.hero-image ul li img {
+    height: 40px;
+}
+
+.hero-image ul li img:hover {
+    transform: scale(1.05);
+    transition: all 0.3s ease-in-out;
+}
+
+/* <!-- ABOUT SECTION --> */
+
+.about {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 20px;
+    margin: 30px 0;
+}
+
+.about h1 {
+    font-size: 45px;
+    font-weight: 500;
+}
+
+.about-text {
+    display: grid;
+    grid-template-columns: 10% 80%;
+    gap: 2%;
+    align-items: center;
+    justify-content: center;
+    padding: 20px 80px;
+}
+
+.about-text hr {
+    border: 1px solid #58287F;
+    width: 100%;
+}
+
+.about-text p span {
+    font-weight: 500;
+    color: #58287F;
+}
+
+/* <-- SKILLS SECTION --> */
+
+.skills {
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+}
+
+.skills h1 {
+    font-size: 45px;
+    font-weight: 500;
+    text-align: center;
+}
+
+.skills-container {
+    display: grid;
+    align-items: center;
+    grid-template-columns: repeat(4, 1fr);
+    padding: 20px 80px;
+    gap: 20px;
+}
+
+.skill-box {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 10px;
+    padding: 60px;
+    min-width: 100px;
+    border-radius: 10px;
+    background-color: rgba(196, 196, 212, 0.25);
+}
+
+.skill-box:hover {
+    cursor: pointer;
+    transform: scale(1.05);
+    transition: all 0.3s ease-in-out;
+}
+
+.skill-box img {
+    height: 80px;
+}
+
+.skill-box p {
+    font-size: 20px;
+    text-transform: uppercase;
+    font-weight: 500;
+    text-align: center;
+}
+
+/* <-- PORTFOLIO --> */
+
+.portfolio {
+    padding: 20px 80px;
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+}
+
+.portfolio h1 {
+    font-size: 45px;
+    font-weight: 500;
+    text-align: center;
+}
+
+.portfolio-container {
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    gap: 20px;
+}
+
+.portfolio-box {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 20px;
+    padding: 20px;
+    text-align: center;
+    border-radius: 10px;
+    width: 80%;
+    margin: 0 auto;
+    padding: 10px 30px;
+
+    /* background-image: url('./images/hero-cover.PNG'); */
+    background-position: center;
+    background-size: cover;
+
+    border: 1px solid #58287F;
+}
+
+.portfolio-box h1 {
+    font-size: 25px;
+    font-weight: 600;
+    text-transform: uppercase;
+}
+
+.portfolio-box p {
+    font-size: 18px;
+    font-weight: 400;
+    word-wrap: break-word;
+    max-width: 350px;
+}
+
+.portfolio-cta {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+}
+
+.portfolio-button {
+    display: flex;
+    justify-content: center;
+
+    padding: 20px;
+}
+
+/* <--- EXPERIENCE SECTION --> */
+
+.experience {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 20px;
+    width: 100%;
+    justify-content: center;
+    padding: 20px 80px;
+}
+
+.experience h1 {
+    font-size: 45px;
+    font-weight: 500;
+    text-align: center;
+}
+
+.experience-container {
+    display: flex;
+    flex-direction: column;
+    gap: 30px;
+
+}
+
+.experience-box {
+    display: grid;
+    grid-template-columns: 10% 80%;
+    align-items: flex-start;
+    gap: 20px;
+    background-color: rgba(196, 196, 212, 0.25);
+    padding: 20px;
+
+    border-radius: 10px;
+}
+
+.experience-logo {
+    height: 100px;
+}
+
+.experience-logo img {
+    height: 100%;
+}
+
+.experience-description {
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+}
+
+.experience-header {
+    height: auto;
+    padding: 10px;
+}
+
+.experience-header h1 {
+    font-size: 25px;
+    font-weight: 500;
+}
+
+.experience-header h2 {
+    font-size: 18px;
+    font-weight: 400;
+    color: #0B9444;
+}
+
+.experience-header p {
+    text-transform: uppercase;
+    font-size: 14px;
+}
+
+.experience-body {
+    height: auto;
+    padding: 10px;
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+}
+
+.experience-body li {
+    font-size: 16px;
+    font-weight: 400;
+}
+
+/* <-- BLOGS SECTION --> */
+
+.blogs {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 20px;
+    padding: 20px 80px;
+}
+
+.blogs h1 {
+    font-size: 45px;
+    font-weight: 500;
+    text-align: center;
+}
+
+.blogs-container {
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    justify-content: center;
+    align-items: center;
+    place-content: center;
+}
+
+.blog-box {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 20px;
+    padding: 0 0 20px 0;
+    text-align: center;
+    border-radius: 10px;
+    border: 1px solid #58287F;
+    max-width: 90%;
+    margin: 0 auto;
+    /* background-color: rgba(196, 196, 212, 0.25); */
+}
+
+.blog-box img {
+    height: 200px;
+    width: 100%;
+    object-fit: cover;
+    border-radius: 10px;
+}
+
+.blog-text h1 {
+    font-size: 20px;
+    padding: 10px 20px;
+    font-weight: 600;
+    text-transform: uppercase;
+}
+
+.blog-box p {
+    font-size: 16px;
+    font-weight: 400;
+    word-wrap: break-word;
+    max-width: 350px;
+    padding: 5px 20px;
+    margin: 0 auto;
+}
+
+.blog-cta {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+}
+
+.blog-cta a {
+    padding: 5px 15px;
+    background-color: #58287F;
+    color: white;
+    text-decoration: none;
+    border-radius: 5px;
+}
+
+.blogs-cta {
+    display: flex;
+    justify-content: center;
+    padding: 20px;
+}
